# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Banco de dados UFS",
        "description": "API Simples Banco de Dados",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:8000/api",
            "description": "SwaggerHUB API Mocking"
        }
    ],
    "paths": {
        "/user/index": {
            "get": {
                "summary": "Retorna todos os usuários cadastrados se existirem",
                "operationId": "0dd4967378edc8e46e723d366584aefb",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "nome": {
                                                "type": "string"
                                            },
                                            "cpf": {
                                                "type": "string"
                                            },
                                            "data_nascimento": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "summary": "Adiciona um usuário ao sistema",
                "operationId": "9039ea5214f8036d1cc688e25b749717",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "type": "string"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "type": "string"
                                        },
                                        "cpf": {
                                            "type": "string"
                                        },
                                        "data_nascimento": {
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}